feat: expose multiple constructors
